Refactor code: Fix memory leak, Remove compile warning  and Replace dynamic link with static link. (#317)
* Use static link


* Add dependency to solve link fail


* Remove compile warning


* Remove more warning


* Fix heap-use-after-free FragmentContext, fix test_insert.py failure (#315)

* Fix heap-use-after-free FragmentContext

* Fix test_insert.py

* Remove value warning


* Fix thrift server memory leak


---------

+++ benchmark/wal/asio_wal.cpp
@@ -393,8 +393,8 @@ main(int argc, char* argv[]) {
         return 1;
     }
 
-    std::int32_t sessions = std::atoi(argv[1]);
-    std::int32_t ioc_threads = std::atoi(argv[2]);
+    std::size_t sessions = std::atoi(argv[1]);
+    std::size_t ioc_threads = std::atoi(argv[2]);
     std::string wal_path(argv[3]);
 
     initializeLogger();
